+++ source/create-form.jsx
@@ -50,13 +50,14 @@ export default (options = {}) => {
 
     addError = ({ name, error }) => {
       if (this.errors[name] !== error) {
-        this.errors[name] = error;
+        this.errors = Object.assign({}, this.errros, { [name]: error });
         this.props.onErrorsChange(this.errors);
       }
     };
 
     removeError = name => {
       this.errors[name] = null;
+      this.errors = Object.assign({}, this.errros);
       this.props.onErrorsChange(this.errors);
     };