+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Make errors object immutable
 
 ## [0.4.2] - 2018-09-30
 - Prevent unnecessary onErrorsChange calls